initial implementation of ActTotalDuration
ActTotalDuration is a simple plugin for the Advanced Combat Tracker. It
adds a "TotalDuration" field to encounters and combatant views.

This new field differs from the existing duration in two key ways:

 * For combatants view, this tracks the time of the encounter, and not
   the time an individual combatant participated. If a combatant leaves
   combat, its duration field will stop incrementing. However, as long
   as the encounter is tracked by ACT, the total duration will continue
   incrementing.

 * For merged encounters, this tracks the total time spent for all
   encounters. This includes any down time between combat. This is the
   main feature of TotalDuration and is what sets it apart from the
   existing duration field.

The TotalDuration field can be useful to help track how long a full
dungeon, raid, or similar takes to complete.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>