remove previous TotalDuration during init
We don't remove the TotalDuration field during DeInitPlugin. Because of
this, if the plugin is removed and re-initialized an error can be
generated during the init process. Avoid this by first removing the
TotalDuration field if it exists. This ensures that re-installations
with a running copy of ACT do not fail.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>