store total number of seconds in the SQL export
For the SQL export, store the total seconds as a double value, rather
than storing it with the format used for the UI display. This is more
useful for other kinds of data processing.

I considered storing it as an integer, rounding the sub-seconds, or as
milliseconds. However, given that TotalSeconds from the TimeSpan is
already a double this made the most sense.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>